Extract process out piping from sg into lib/process
